Set large StreamReadConstraints.maxStringLength for IR parsing only
Enlarge the maxStringLength just before parsing the (perhaps huge!) IR
and reset it to the default immediately afterwards.

If this code is multithreaded and other threads also parse JSON, there
is a race condition here. It would be better to construct a bespoke
JsonFactory to parse this IR, but expressing that while satisfying the
type checker requires more Scala knowledge than we have.